blockchain_manager/pyproject.toml
[tool.poetry]
name = "blockchain-manager"
version = "0.1.0"
description = ""
authors = ["Santiago Vasquez Gomez <santiago.vasquez@globalmvm.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
kubernetes = "^32.0.0"
httpx = "^0.28.1"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
azure-functions = "^1.21.3"
global-pydependency = {path = "global_pydependency/global_pydependency-0.1.0.tar.gz"}


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
